Reorganize api tests into versioned structure
