+++ tests/Feature/API/V1/TestCase.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Feature\API\V1;
+
+class TestCase extends \Tests\TestCase
+{
+    public function apiUrl($path): string
+    {
+        return "/api/v1/$path";
+    }
+}
